+++ Analysis.h
@@ -129,16 +129,6 @@ private:
      double _cutRadius;
      int _circleEdges;
      bool _getProfile;        // Whether make field analysis or not
-     bool _outputGraph;       // Whether output the data for plot the voronoi diagram each frame
-     bool _plotGraph;       // Whether plot the voronoi diagram each
-                            // frame. if (outputGraph==true)
-     bool _plotIndex;       // Whether plot the voronoi diagram each
-                            // frame with index of pedesrians
-                            // if (outputGraph==true and _polotGraph==true)
-     std::vector<bool> _plotTimeseriesA;
-     std::vector<bool> _plotTimeseriesC;
-     std::vector<bool> _plotTimeseriesD;
-     std::vector<bool> _plotTimeseriesI;
      bool _isOneDimensional;
      bool _calcIndividualFD;  //Adjust whether analyze the individual density and velocity of each pedestrian in stationary state (ALWAYS VORONOI-BASED)
      std::string _vComponent;        // to mark whether x, y or x and y coordinate are used when calculating the velocity